refactor(core): extract SDK components into separate modules
Reduced main SDK file from 1227 to 587 lines (52% reduction) by extracting:

- **Engine Asset Loading** (engine_asset_loader.dart):
  - Asset loading from local storage and bundle
  - Android asset path fixing
  - Vendor asset injection (jQuery, etc.)
  - Alternative JS file name resolution

- **Handler Factory** (handler_factory.dart):
  - Required handlers creation for FormGear and FasihForm
  - Assignment-aware handlers with context
  - Centralized handler configuration

- **WebView Builder** (webview_builder.dart):
  - WebView creation with assignment context
  - FormGearEnginePage wrapper widget
  - JavaScript executor service integration

Benefits:
- Improved code organization and maintainability
- Better separation of concerns
- Easier testing of individual components
- Reduced complexity in main SDK class
